# lib/api/v1/hosonto_api.js
var util = require("util");
var config = require("../../config");
var log = require("../../log")(module);
var serverTemplates = require("../../messages");

module.exports = function (app, myConn) {
  let deserialize = function (str, separation, array) {
    var index = function (obj, i) {
      return obj[i];
    };
    return str.split(separation).reduce(index, array);
  };

  let sendHttpResponse = function (res, code, message) {
    errorMessage = serverTemplates.http_messages[code];
    if (code === 200) {
      res.status(200).json(message);
    } else {
      if (config.get("LOG_LEVEL") === "debug") {
        errorMessage = message;
      }
      log.debug("\n\n\nServer error: " + code + " " + JSON.stringify(message));
      res.status(code || 500).json(errorMessage);
    }
  };

  let isAutoDataAllowed = function (req, res, next) {
    log.debug(
      "\n\nAuto Data Api not permitted for security reason, use executeEvent and event endpoint instead."
    );
    if (config.get("ALLOW_AUTODATAHANDLER") === true) {
      return next();
    }
    sendHttpResponse(res, 403);
  };

  let fArray = Array.from(
    Array(config.get("MAX_FIELD_IN_HTTP_REQ")),
    (x, ind) => "/:fieldName" + ind + "?/:fieldValue" + ind + "?"
  );

  app.get(
    "/autoLoadData/:tableName" + fArray.join(""),
    isAutoDataAllowed,
    async function (req, res) {
      log.debug(
        JSON.stringify(req.body) + " params:" + JSON.stringify(req.params)
      );
      let tableName = req.params["tableName"] + "_TABLE";
      let params = {};
      let searchQry = {};
      params[tableName] = [];
      searchQry[req.params["fieldName"] || req.query["fieldName"]] =
        req.params["fieldValue"] || req.query["fieldValue"];
      for (let i = 0; i < 100; i++) {
        searchQry[req.params["fieldName" + i] || req.query["fieldName" + i]] =
          req.params["fieldValue" + i] || req.query["fieldValue" + i];
      }
      params[req.params["tableName"] + "_FIELD_"] = JSON.parse(
        JSON.stringify(searchQry)
      );
      try {
        let loadAutoTablePromise = util.promisify(myConn.loadAutoTable);
        params = await loadAutoTablePromise(
          params /*,
          function callback(err, sessionId) {
            params["__SessionId"] = sessionId;
            global._params = params;
            log.debug("LoadTable Params=" + util.inspect(params));
            sendHttpResponse(res, (err && err.code) || 200, params);
          }*/
        );
        log.debug("Returned params" + JSON.stringify(params));
        sendHttpResponse(res, 200, params);
      } catch (error) {
        log.debug("Returned error" + JSON.stringify(error));
        sendHttpResponse(res, error & error.code, error);
      }
    }
  );

  app.post("/autoLoadData", isAutoDataAllowed, function (req, res) {
    try {
      var params = req.body;

      params["session"] = req.session;
      if (req.session && req.session.passport) {
        params["session_user"] = req.session.passport.user;
      }

      log.debug(
        "\n\nsession::" +
          JSON.stringify(req.session) +
          " &&" +
          params["session_user"] +
          ":::" +
          params["__SessionId"]
      );

      myConn.loadAutoTable(params, function callback(err, sessionId) {
        params["__SessionId"] = sessionId;
        global._params = params;
        log.trace("LoadTable Params=" + util.inspect(params));
        sendHttpResponse(res, (err && err.code) || 200, params);
      });
    } catch (error) {
      sendHttpResponse(res, 500, error);
    }
  });

  let callFunction = async function (req, res, params, fn) {
    try {
      let passed = await fn(params, req, res);

      if (params["__UseServerSideFunction"] === true) {
        log.trace(
          "\n Returning by server side execution" + util.inspect(passed)
        );
        sendHttpResponse(res, 200, passed);
      } else {
        log.trace("\n Before save params: " + util.inspect(passed));
        myConn.saveAutoParams(passed, (err, params) => {
          log.trace("\n executeEvent::  Loading params" + util.inspect(params));
          myConn.loadAutoParams(params, (err, params) => {
            log.trace(
              "\n\n\n After Load Params" +
                "\n" +
                JSON.stringify(err) +
                "\n" +
                util.inspect(params)
            );
            if (!err) {
              let retParams = { ...params };
              Object.keys(params).forEach((key) => {
                if (key.includes("_TABLE")) {
                  let [tableName, desc] = key.split("_TABLE");
                  delete retParams[tableName + "__search"];
                  delete retParams[tableName + "__sort"];
                  delete retParams[tableName + "__bi"];
                }
              });
              sendHttpResponse(res, 200, retParams);
            } else {
              sendHttpResponse(res, err && err.code, []);
            }
          });
        });
      }
    } catch (err) {
      throw err;
    }
  };

  app.post("/executeEvent", function (req, res) {
    var params = req.body;

    log.debug(
      " Request body :: " +
        JSON.stringify(req.body) +
        req.body +
        " request params:" +
        JSON.stringify(req.params)
    );

    params["__IP"] =
      req.ip ||
      (
        (req.headers["x-forwarded-for"] || "").split(",").pop() ||
        req.connection.remoteAddress ||
        req.socket.remoteAddress ||
        req.connection.socket.remoteAddress
      )
        .split(":")
        .slice(-1);

    params["session"] = req.session;
    if (req.session && req.session.passport) {
      params["session_user"] = req.session.passport.user;
    }

    var sessionId = params["__SessionId"];

    var fn = deserialize(params["__EventName"], ".", global);

    global._params = params;
    log.debug(
      "\n Execute fn: " + fn + "\n with Params: " + util.inspect(params)
    );

    try {
      let { isAuthorized, role } = req.isAuthenticated();
      if (
        (fn.allowGlobal && fn.allowGlobal == true) ||
        (isAuthorized && fn.authorizationLevel >= role)
      ) {
        try {
          callFunction(req, res, params, fn);
        } catch (error) {
          sendHttpResponse(res, error && error.code);
        }
      } else if (!isAuthorized) {
        let newParams = {};
        newParams["__Error"] = serverTemplates.auth_messages.login_required;
        myConn.saveAutoParams(newParams, (err, params) => {
          myConn.loadAutoParams(params, (err, params) => {
            if (!err) {
              sendHttpResponse(res, 403, params);
            } else {
              sendHttpResponse(res, err && err.code, params);
            }
          });
        });
      }
    } catch (error) {
      console.log("\n\n\n Err occurred" + JSON.stringify(error));
      sendHttpResponse(res, 500, error);
    }
  });

  let getRemoteFunction = function (req, res, fnName, params) {
    var fn = deserialize(fnName, ".", global);

    log.debug("Found fn= " + fn + " \nReq=" + req + "\nParams=" + params);

    try {
      if (!fn) {
        params["__Error"] = serverTemplates.server_messages.remote_fn_not_found;
        myConn.saveAutoParams(params, () => {
          // ((res, params) => {
          // return function callback() {
          myConn.loadAutoParams(params, (err, results) => {
            // ((res, params) => {
            // return function callback(err, results) {
            sendHttpResponse(res, 404, { __Error: params.__Error });
            // };
            // })(res, params)
          });
          // };
          // })(res, params)
        });
      } else if (fn.allowRemote && fn.allowRemote == true) {
        strFn = fn + " ";
        params["__RemoteFunction"] = strFn;
        log.debug("In param=" + params["__RemoteFunction"]);
        myConn.saveAutoParams(params, () => {
          // ((res, params) => {
          // return function callback() {
          myConn.loadAutoParams(params, (err, results) => {
            // ((res, params) => {
            // return function callback(err, results) {
            sendHttpResponse(res, 200, {
              __RemoteFunction: params.__RemoteFunction,
            });
            // };
            // })(res, params)
          });
          // };
          // })(res, params)
        });
      } else {
        params["__Error"] = serverTemplates.remote_fn_not_permitted;
        myConn.saveAutoParams(params, () => {
          // ((res, params) => {
          // return function callback() {
          myConn.loadAutoParams(params, (err, results) => {
            // (function (res, params) {
            // return function callback(err, results) {
            sendHttpResponse(res, 403, { __Error: params.__Error });
            // };
            // })(res, params)
          });
          // };
          // })(res, params)
        });
      }
    } catch (error) {
      sendHttpResponse(res, 500, error);
    }
  };

  app.get("/getRemoteFunction/:fnName", function (req, res) {
    var params = req.params;
    let fnName = req.params["fnName"];
    params["__IP"] =
      req.ip ||
      (
        (req.headers["x-forwarded-for"] || "").split(",").pop() ||
        req.connection.remoteAddress ||
        req.socket.remoteAddress ||
        req.connection.socket.remoteAddress
      )
        .split(":")
        .slice(-1);

    params["session"] = req.session;
    if (req.session && req.session.passport) {
      params["session_user"] = req.session.passport.user;
    }

    getRemoteFunction(req, res, fnName, params);
  });

  app.post("/getRemoteFunction", function (req, res) {
    var params = req.body;

    params["__IP"] =
      req.ip ||
      (
        (req.headers["x-forwarded-for"] || "").split(",").pop() ||
        req.connection.remoteAddress ||
        req.socket.remoteAddress ||
        req.connection.socket.remoteAddress
      )
        .split(":")
        .slice(-1);

    params["session"] = req.session;
    if (req.session && req.session.passport) {
      params["session_user"] = req.session.passport.user;
    }

    var sessionId = params["__SessionId"];

    getRemoteFunction(req, res, params["__EventName"], params);
  });
};